# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Habit Backend API",
        "description": "API de gestion de boutiques de vêtements — stocks, ventes, caisse, rapports.",
        "contact": {
            "email": "admin@shop.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/rapports/ventes": {
            "get": {
                "tags": [
                    "Rapports"
                ],
                "summary": "Ventes groupées par période",
                "operationId": "af169f27d6514a8fc2004810218f0d19",
                "parameters": [
                    {
                        "name": "groupBy",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "jour",
                            "enum": [
                                "jour",
                                "semaine",
                                "mois"
                            ]
                        }
                    },
                    {
                        "name": "dateDebut",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "dateFin",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "boutiqueId",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ventes par période",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rapports/resume-dashboard": {
            "get": {
                "tags": [
                    "Rapports"
                ],
                "summary": "Résumé tableau de bord",
                "operationId": "9774e84c12ad6fbc403bd0defee839e2",
                "parameters": [
                    {
                        "name": "boutiqueId",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "ventesAujourdhui": {
                                                    "type": "string"
                                                },
                                                "alertesStock": {
                                                    "type": "integer"
                                                },
                                                "produitsActifs": {
                                                    "type": "integer"
                                                },
                                                "topProduits": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "nullable": true
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rapports/export-excel": {
            "get": {
                "tags": [
                    "Rapports"
                ],
                "summary": "Export des ventes en Excel (.xlsx)",
                "operationId": "fc082e068cc8d9dd7c850879ce01dc6f",
                "parameters": [
                    {
                        "name": "dateDebut",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "dateFin",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "boutiqueId",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier Excel",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rapports/export-pdf": {
            "get": {
                "tags": [
                    "Rapports"
                ],
                "summary": "Export des ventes en PDF",
                "operationId": "ca8be829b4e7ec4e5cf80e93604cd5bb",
                "parameters": [
                    {
                        "name": "dateDebut",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "dateFin",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "boutiqueId",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier PDF",
                        "content": {
                            "application/pdf": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}